+++ packages/domain-lib/test/unit/aggregate_events_non_happy_path.test.ts
